Revert "feat: enable animation of more point style properties"
This reverts commit c05f52066e4120804ccb25c4372a8e088c668c28.
